refactor: creating openapi.yaml

+++ internal/posts/postcomments/repository.go
@@ -9,9 +9,10 @@ import (
 	"github.com/manuelarte/gowasp/internal/models"
 )
 
+//nolint:iface // separate repository from service
 type Repository interface {
 	Create(ctx context.Context, postComment *models.PostComment) error
-	GetAllForPostID(ctx context.Context, postID uint64,
+	GetAllForPostID(ctx context.Context, postID uint,
 		pageRequest *pagorminator.Pagination) ([]*models.PostComment, error)
 }
 
@@ -25,7 +26,7 @@ func NewRepository(db *gorm.DB) Repository {
 	return &gormRepository{db: db}
 }
 
-func (b gormRepository) GetAllForPostID(ctx context.Context, postID uint64,
+func (b gormRepository) GetAllForPostID(ctx context.Context, postID uint,
 	pageRequest *pagorminator.Pagination,
 ) ([]*models.PostComment, error) {
 	var postComments []*models.PostComment
